# docs/api/openapi/control-center.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgainERP Control Center API",
    "description": "\nCentral control plane REST API for AgainERP operators and Edge Agents.\n\n**Operator API** — prefix `/api/v1/*` — requires `Authorization: Bearer {jwt}` unless noted.\n\n**Agent API** — prefix `/agent/v1/*` — requires agent bearer token (mTLS in production).\n\n**Webhooks** — prefix `/webhooks/v1/*` — provider-signed inbound events.\n\nHigh-risk actions (e.g. registration approval) require MFA step-up within a 5-minute window when MFA is enabled.\n",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/operators/": {
      "get": {
        "tags": [
          "operators"
        ],
        "summary": "List Operators",
        "operationId": "list_operators_api_v1_operators__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {},
                  "type": "array",
                  "title": "Response List Operators Api V1 Operators  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/operators/me": {
      "get": {
        "tags": [
          "operators"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_v1_operators_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Me Api V1 Operators Me Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/licenses/": {
      "get": {
        "tags": [
          "licenses"
        ],
        "summary": "List Licenses",
        "operationId": "list_licenses_api_v1_licenses__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Client Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {},
                  "title": "Response List Licenses Api V1 Licenses  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "licenses"
        ],
        "summary": "Create License Route",
        "operationId": "create_license_route_api_v1_licenses__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LicenseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create License Route Api V1 Licenses  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/licenses/validate": {
      "post": {
        "tags": [
          "licenses"
        ],
        "summary": "Validate License Route",
        "operationId": "validate_license_route_api_v1_licenses_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LicenseValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Validate License Route Api V1 Licenses Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/licenses/{license_id}": {
      "get": {
        "tags": [
          "licenses"
        ],
        "summary": "Get License",
        "operationId": "get_license_api_v1_licenses__license_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "license_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "License Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get License Api V1 Licenses  License Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LicenseCreate": {
        "properties": {
          "client_id": {
            "type": "string",
            "title": "Client Id"
          },
          "plan": {
            "type": "string",
            "title": "Plan",
            "default": "starter"
          },
          "subscription_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription Id"
          },
          "expires_days": {
            "type": "integer",
            "title": "Expires Days",
            "default": 365
          }
        },
        "type": "object",
        "required": [
          "client_id"
        ],
        "title": "LicenseCreate"
      },
      "LicenseValidateRequest": {
        "properties": {
          "license_key": {
            "type": "string",
            "title": "License Key"
          }
        },
        "type": "object",
        "required": [
          "license_key"
        ],
        "title": "LicenseValidateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}